+++ mc26.1/common/src/main/java/com/bondigi/seedblend/chunk/SeedBlendChunkDataAccess.java
@@ -18,4 +18,8 @@ public interface SeedBlendChunkDataAccess {
 
     @Nullable
     DimensionBlendPolicy seedblend$getPolicy();
+
+    int seedblend$getTransitionWeight();
+
+    void seedblend$setTransitionWeight(int weightPercent);
 }